Make sure the Swagger UI leverages our already-defined OpenAPI files
It seems rather odd that the Swagger UI interfaces exposes
endpoints different from the definition files.
Instead, SpringFox attempts to deduce what to expose from the SpringBoot controllers.

To further improve the API-first approach, we instead instruct SpringFox
to leverage the OpenAPI definitions we already have on file.
TO do this however, an OpenAPI JSON file is needed as input, which is the goal
of the 'convertOpenAPIYaml2Json' Gradle task added.