No longer generate Swagger Annotations
Swagger Annotions are useful in a code-first approach,
in which SpringFox is able to generate the OpenAPI definition
from the code and expose the Swagger UI based upon such code
annotations.

But, we are now leveraging an API-first driven approach instead.
So everything (including the Swagger UI) is governed by the
static OpenAPI definition files we have.