Expose the OpenAPI YAML file and infer the Swagger UI base URL from the root Context Path
This operation needs to be done at runtime, hence the need for a dedicated endpoint.
As a consequence, conversion of the OpenAPI YAML file to a JSON file
is no longer done at build time, but at runtime in this new endpoint.
For performance purposes, parsing is done only once, lazily.