[api] Make the 'copySubProjectsOpenAPIFiles' task read its inputs from the actual project dependencies rather than the parent subprojects
This prevents exposing endpoints in the Swagger UI (since the project OpenAPI file gets merged),
but without the actual code dependency, resulting in a 404 error returned.
This makes it clear that an explicit dependency to the project has to be declared,
for Spring to auto-discover the service and controller beans.