Fix issue preventing examples for request bodies from rendering properly in the Swagger UI
This caused the following JavaScript issue:
```js
Could not render n, see the console.
```

This is because cloning the base parsed OpenAPI object via the ObjectMapper
used to add additional invalid fields (like a 'oneOf' field in schemas with the original schema, but a new one of type 'array' but with no 'items' declared).

A copy was being performed to alter the Version and base servers URLs.
But with hindsight, this possibly-expensive operation is not needed,
as the parsed OpenAPI is actually mutable.