Inject custom HttpMessageConverter that can read and write YAML
This will make it easier to (de-)serialize data based on the
MediaType, thus removing the need for a manual handling code.

As an example, we could use the same `@RequestBody` annotation
in a controller method as long as 'application/yaml' is registered
as an Acceptable header.
The following media types are supported:
- application/yaml
- application/yml
- text/yaml
- text/yml
- application/*+yaml
- application/*+yml