Pass imagePullSecrets to Argo Workflows Pipelines as needed
The main reason for this is, when deploying through the Azure
Marketplace, we provision an Azure Container Registry (ACR) in
the target Tenant so they can push their own Solution images
as needed. In turn, the Kubernetes Kubelets need to be able to
pull such private images.
Attaching the ACR to AKS ended up being not that reliable,
because of the need to wait a lot of minutes,
with no guarantee whatsoever that the permissions wil be propagated.
Kubernetes ImagePullSecrets are a faster way to fix this issue.

This introduces the following properties:
- csm.platform.argo.image-pull-secrets : a list of Docker credentials
secrets to set in all Argo Workflows. Must exist in the Kubernetes Cluster.

Note that the Secret and the Configuration are auto-created when deploying
via our Helm Chart, as long as the following values are provided at
deployment time:
- argo.imageCredentials.registry
- argo.imageCredentials.username
- argo.imageCredentials.password