chore: Ignore "CIS-DI-0006" container scan best practice check
CIS-DI-0006 is raised because the resulting image has no HEALTHCHECK instruction.

HEALTHCHECK is not part of the OCI spec, but is very Docker-specific.
And, as commented in [1], Jib has no intention to support it,
unless the core OCI spec supports it.

Plus, Kubernetes makes use of its own Startup / Liveness / Readiness
probes for health checks.

[1] https://github.com/GoogleContainerTools/jib/issues/676#issuecomment-442533758