[PROD-7730] Extract all calls to Argo in a single Service bean
Having too many components manipulating Argo makes it harder
to understand the logic behind scenario runs, and also makes it
harder to refactor the code if needed.

Calls to Argo are an implementation detail (for running scenarios),
which should not leak into the other services.

This would allow to make this new Service bean participate
to the Application Readiness State.