Tolerate a slight service disruption of 'argo-minio' when updating it
As Minio uses a ReadWriteOnce PVC by default, using a RollingUpdate strategy will
not work if the new pod is scheduled on a different node.
It is possible to force the replicas scheduling on a same node, but this requires
that node to have a minimum of 4Gi additional memory available,
which is the default memory request set by Argo.