Relax if an error occurs while fetching Workflows statuses
Just return null, which indicates an unknown status.

## Context

Workflows IDs are specified to a given environment.
When making a local API point to a given remote environment,
it becomes impossible to locally list scenarios, as an attempt is
made to contact Argo to get the Workflow status.
But the local Argo server has no knowledge about that scenario run workflow id.