feat: Limit the total number of workflows, pods and PVCs handled by Argo
A workflow (and actually, any Kubernetes resource) will incur a cost
as long as they exist in the cluster, even after they are no longer running.
As depicted in the Cost Optimization recommendations [1], we should:

- make sure workflows do not run forever. Hard limit is arbitrarily set to a
  hopefully large value of 7 days (604800 seconds)
- delete successful workflows after 1 day. We can still view them since they are archived
- schedule pods and PVCs for immediate deletion as soon as workflows are successful.
  This should fix the quota errors that might be returned by Azure when leveraging the
  AzureFile storage provisioner, by wiping out successful workflow data.
  We can still access their logs and artifacts since they are archived.

These represent default values applicable to any Workflow handled by the
Argo Workflow Controller, but such values can be overridden as needed on a
per-Workflow basis.

[1] https://argoproj.github.io/argo-workflows/cost-optimisation/#limit-the-total-number-of-workflows-and-pods