chore(CD): Use a repository_dispatch event to trigger the 'deploy' workflow
The 'workflow_run' event does not support passing extra information,
like the exact SHA of the origin Workflow. [1]
This is an issue because if other changes were pushed before the 'deploy'
workflow starts, the 'checkout' Action would pull the latest HEAD,
and we might not deploy the right changes.
Worse, the 'deploy' Workflow might not pass if
the container images and Helm charts were not yet pushed.

[1] https://github.community/t/pass-data-from-original-workflow-that-triggered-current-workflow/148988