[PROD-7420] Instruct the engine to push ScenarioRun related data to the Control Plane Event Bus
This is a prerequisite for implementing this,
as we will base the data ingestion state computation on this control plane Event Bus.

In a nutshell:
- a new SimulationTotalFacts table will be added to each Azure Data Explorer Database,
along with a data connection to this control plane Event Hub and a dedicated ingestion mapping.
This new table will contain the total number of messages and facts expected for each Simulation.
This will be populated at the end of each Simulation.
- when a request for a scenario is made, we already issue an additional request to the
Workflow Service to determine the workflow execution status. If completed, we will then issue
an additional query to Azure Data Explorer to count the number of probe measures vs the total
number of messages expected.

Edge cases to consider:
- what if the control plane message never arrives or there is some kind of ingestion latency ?
- what happens if we have no consumers? This should however be fixed at some point in the Simulator engine,
by making sure a message is sent with 0 as total messages expected.
- what about backward compatibility for older simulator engines. If there is nothing in SimulationTotalFacts,
but there exists data in the probe measures table?

These will hopefully be addressed in the upcoming commits.

For reference, here is the Kusto query for creating this new database table:

```
.create table SimulationTotalFacts(
SimulationId:string,
State:string,
Type:string,
SentMessagesTotal: long,
SentFactsTotal: long)

.create table SimulationTotalFacts ingestion json mapping "SimulationTotalFactsMapping"
    '['
    '    { "column" : "SimulationId", "Properties":{"Path":"$.sagaId"}},'
    '    { "column" : "State", "Properties":{"Path":"$.state"}},'
    '    { "column" : "Type", "Properties":{"Path":"$.type"}},'
    '    { "column" : "SentMessagesTotal", "Properties":{"Path":"$.sentMessagesTotal"}},'
    '    { "column" : "SentFactsTotal", "Properties":{"Path":"$.sentFactsTotal"}},'
    ']'
```