Be liberal in what we accept as 'computeSize'
In other words, it does not matter if 'computeSize' ends with the 'pool' suffix.
Instead, we should sanitize the value to conform to the format we expect.
Otherwise, the Kubernetes cluster might end up with lots of Pending Pods,
because of a non-conformant node selector,
indefinitely waiting for nodes that will never get provisioned.